Rename Tab components, add TabPanel
