+++ services/main-frontend/src/components/TabLinkPanel.tsx
@@ -1,0 +1,9 @@
+const TabLinkPanel: React.FC = ({ children }) => {
+  return (
+    <div role={"tabpanel"} tabIndex={0}>
+      {children}
+    </div>
+  )
+}
+
+export default TabLinkPanel
