+++ services/main-frontend/src/components/page-specific/manage/courses/id/feedback/CourseFeedback.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from "react-i18next"
 
 import createUnreadFeedbackCountHook from "../../../../../../hooks/count/useUnreadFeedbackCount"
 import { CourseManagementPagesProps } from "../../../../../../pages/manage/courses/[id]/[...path]"
-import TabLink from "../../../../../LinkTab"
-import TabLinkNavigation from "../../../../../LinkTabNavigation"
+import TabLink from "../../../../../TabLink"
+import TabLinkNavigation from "../../../../../TabLinkNavigation"
 
 import FeedbackList from "./FeedbackList"
